Fix forecast scaling to use precise day-based calculations
Previously, forecasts were bucketed into 1/3/12 months, causing
"Next 30 Days" and "Next Month" to show identical values.

Now uses actual day count divided by 30.44 (average days per month)
for precise scaling:
- Next 30 Days (30 days) = 0.99 months
- Next Month Feb (28 days) = 0.92 months
- Next 90 Days = 2.96 months
- Next Quarter = ~3 months
- Next 365 Days = 11.99 months

+++ ArgoBooks.Core/Models/Insights/InsightsData.cs
@@ -88,9 +88,9 @@ public class InsightsSummary
 public class ForecastData
 {
     /// <summary>
-    /// Number of months the forecast period covers (1 = month, 3 = quarter, 12 = year).
+    /// Number of months the forecast period covers (e.g., 0.99 for 30 days, 2.96 for 90 days).
     /// </summary>
-    public int PeriodMonths { get; set; } = 1;
+    public decimal PeriodMonths { get; set; } = 1;
 
     /// <summary>
     /// Forecasted revenue for the next period.
